backend/requirements.txt
# backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-multipart==0.0.6
python-dateutil==2.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# Firestore dependencies (NEW)
google-cloud-firestore==2.13.1
google-cloud-core==2.4.1

# Geolocation dependencies  
cachetools==5.3.2
asyncpg==0.29.0